+++ tests/utils.test.ts
@@ -1,10 +1,5 @@
 import { assertEquals as equals } from "../deps/assert.ts";
-import {
-  isPlainObject,
-  merge,
-  searchByExtension,
-  sha1,
-} from "../core/utils.ts";
+import { isPlainObject, merge, sha1 } from "../core/utils.ts";
 
 Deno.test("merge options", () => {
   interface Options {
@@ -62,23 +57,6 @@ Deno.test("merge inner options", () => {
   };
 
   equals(expected, merge(defaults, user));
-});
-
-Deno.test("search by extension", () => {
-  const extensions = new Map([
-    [".tmpl.ts", "ts template"],
-    [".tmpl.js", "js template"],
-    [".js", "js"],
-    [".ts", "ts"],
-  ]);
-
-  equals(searchByExtension("file.tmpl.ts", extensions), [
-    ".tmpl.ts",
-    "ts template",
-  ]);
-  equals(searchByExtension("file.ts", extensions), [".ts", "ts"]);
-  equals(searchByExtension(".ts", extensions), [".ts", "ts"]);
-  equals(searchByExtension("foo", extensions), undefined);
 });
 
 Deno.test("isPlainObject", () => {